# Conflicts:
#	README.md
#	src/config/CryptoNoteConfig.h
#	src/zedwallet++/CommandImplementations.cpp

+++ src/WalletBackend/WalletBackend.cpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2018-2019, The TurtleCoin Developers
-// 
+//
 // Please see the included LICENSE file for more information.
 
 ////////////////////////////////////////
@@ -33,6 +33,8 @@
 #include <iomanip>
 
 #include "JsonHelper.h"
+
+#include <Logger/Logger.h>
 
 #include <Mnemonics/Mnemonics.h>
 
@@ -97,7 +99,7 @@ Error checkNewWalletFilename(std::string filename)
 
     /* Don't leave random files around if we fail later down the road */
     fs::remove(filename);
-    
+
     return SUCCESS;
 }
 
@@ -112,7 +114,7 @@ WalletBackend::WalletBackend()
 {
     m_eventHandler = std::make_shared<EventHandler>();
 
-    /* Remember to correctly initialize the daemon - 
+    /* Remember to correctly initialize the daemon -
     we can't do it here since we don't have the host/port, and the json
     serialization uses the default constructor */
 }
@@ -137,11 +139,12 @@ WalletBackend::WalletBackend(
     const uint64_t scanHeight,
     const bool newWallet,
     const std::string daemonHost,
-    const uint16_t daemonPort) :
+    const uint16_t daemonPort,
+    const bool daemonSSL) :
 
     m_filename(filename),
     m_password(password),
-    m_daemon(std::make_shared<Nigel>(daemonHost, daemonPort))
+    m_daemon(std::make_shared<Nigel>(daemonHost, daemonPort, daemonSSL))
 {
     /* Generate the address from the two private keys */
     std::string address = Utilities::privateKeysToAddress(
@@ -163,11 +166,12 @@ WalletBackend::WalletBackend(
     const std::string address,
     const uint64_t scanHeight,
     const std::string daemonHost,
-    const uint16_t daemonPort) :
+    const uint16_t daemonPort,
+    const bool daemonSSL) :
 
     m_filename(filename),
     m_password(password),
-    m_daemon(std::make_shared<Nigel>(daemonHost, daemonPort))
+    m_daemon(std::make_shared<Nigel>(daemonHost, daemonPort, daemonSSL))
 {
     bool newWallet = false;
 
@@ -227,12 +231,13 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importWalletFro
     const std::string password,
     const uint64_t scanHeight,
     const std::string daemonHost,
-    const uint16_t daemonPort)
+    const uint16_t daemonPort,
+    const bool daemonSSL)
 {
     /* Check the filename is valid */
     if (Error error = checkNewWalletFilename(filename); error != SUCCESS)
     {
-        return {error, nullptr}; 
+        return {error, nullptr};
     }
 
     /* Convert the mnemonic into a private spend key */
@@ -240,7 +245,7 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importWalletFro
 
     if (mnemonicError)
     {
-        return {mnemonicError, nullptr}; 
+        return {mnemonicError, nullptr};
     }
 
     Crypto::SecretKey privateViewKey;
@@ -250,13 +255,18 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importWalletFro
         privateSpendKey, privateViewKey
     );
 
+    if (Error error = validatePrivateKey(privateViewKey); error != SUCCESS)
+    {
+        return {error, nullptr};
+    }
+
     /* Just defining here so it's more obvious what we're doing in the
        constructor */
     bool newWallet = false;
 
     const std::shared_ptr<WalletBackend> wallet(new WalletBackend(
         filename, password, privateSpendKey, privateViewKey,
-        scanHeight, newWallet, daemonHost, daemonPort
+        scanHeight, newWallet, daemonHost, daemonPort, daemonSSL
     ));
 
     wallet->init();
@@ -276,12 +286,23 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importWalletFro
     const std::string password,
     const uint64_t scanHeight,
     const std::string daemonHost,
-    const uint16_t daemonPort)
+    const uint16_t daemonPort,
+    const bool daemonSSL)
 {
     /* Check the filename is valid */
     if (Error error = checkNewWalletFilename(filename); error != SUCCESS)
     {
-        return {error, nullptr}; 
+        return {error, nullptr};
+    }
+
+    if (Error error = validatePrivateKey(privateViewKey); error != SUCCESS)
+    {
+        return {error, nullptr};
+    }
+
+    if (Error error = validatePrivateKey(privateSpendKey); error != SUCCESS)
+    {
+        return {error, nullptr};
     }
 
     /* Just defining here so it's more obvious what we're doing in the
@@ -290,7 +311,7 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importWalletFro
 
     const std::shared_ptr<WalletBackend> wallet(new WalletBackend(
         filename, password, privateSpendKey, privateViewKey, scanHeight,
-        newWallet, daemonHost, daemonPort
+        newWallet, daemonHost, daemonPort, daemonSSL
     ));
 
     wallet->init();
@@ -310,17 +331,30 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::importViewWalle
     const std::string password,
     const uint64_t scanHeight,
     const std::string daemonHost,
-    const uint16_t daemonPort)
+    const uint16_t daemonPort,
+    const bool daemonSSL)
 {
     /* Check the filename is valid */
     if (Error error = checkNewWalletFilename(filename); error != SUCCESS)
     {
-        return {error, nullptr}; 
+        return {error, nullptr};
+    }
+
+    if (Error error = validatePrivateKey(privateViewKey); error != SUCCESS)
+    {
+        return {error, nullptr};
+    }
+
+    const bool allowIntegratedAddresses = false;
+
+    if (Error error = validateAddresses({address}, allowIntegratedAddresses); error != SUCCESS)
+    {
+        return {error, nullptr};
     }
 
     const std::shared_ptr<WalletBackend> wallet(new WalletBackend(
         filename, password, privateViewKey, address, scanHeight, daemonHost,
-        daemonPort
+        daemonPort, daemonSSL
     ));
 
     wallet->init();
@@ -336,14 +370,15 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::createWallet(
     const std::string filename,
     const std::string password,
     const std::string daemonHost,
-    const uint16_t daemonPort)
+    const uint16_t daemonPort,
+    const bool daemonSSL)
 {
     /* Check the filename is valid */
     if (Error error = checkNewWalletFilename(filename); error != SUCCESS)
     {
         return {error, nullptr};
     }
-	
+
     CryptoNote::KeyPair spendKey;
     Crypto::SecretKey privateViewKey;
     Crypto::PublicKey publicViewKey;
@@ -363,9 +398,9 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::createWallet(
 
     const std::shared_ptr<WalletBackend> wallet(new WalletBackend(
         filename, password, spendKey.secretKey, privateViewKey,
-        scanHeight, newWallet, daemonHost, daemonPort
+        scanHeight, newWallet, daemonHost, daemonPort, daemonSSL
     ));
-	
+
     wallet->init();
 
     /* Save to disk */
@@ -379,7 +414,8 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::openWallet(
     const std::string filename,
     const std::string password,
     const std::string daemonHost,
-    const uint16_t daemonPort)
+    const uint16_t daemonPort,
+    const bool daemonSSL)
 {
     /* Open in binary mode, since we have encrypted data */
     std::ifstream file(filename, std::ios_base::binary);
@@ -387,7 +423,7 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::openWallet(
     /* Check we successfully opened the file */
     if (!file)
     {
-        return {FILENAME_NON_EXISTENT, nullptr}; 
+        return {FILENAME_NON_EXISTENT, nullptr};
     }
 
     /* Read file into a buffer */
@@ -403,7 +439,7 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::openWallet(
 
     if (error)
     {
-        return {error, nullptr}; 
+        return {error, nullptr};
     }
 
     using namespace CryptoPP;
@@ -467,7 +503,7 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::openWallet(
 
     if (error)
     {
-        return {error, nullptr}; 
+        return {error, nullptr};
     }
 
     try
@@ -494,13 +530,19 @@ std::tuple<Error, std::shared_ptr<WalletBackend>> WalletBackend::openWallet(
         /* Initialize it from the json (We could do this in less steps, but it
            requires a move/copy constructor) */
         error = wallet->fromJSON(
-            walletJson, filename, password, daemonHost, daemonPort
+            walletJson, filename, password, daemonHost, daemonPort, daemonSSL
         );
 
         return {error, wallet};
     }
     catch (const std::invalid_argument &e)
     {
+        Logger::logger.log(
+            std::string("Failed to open wallet file: ") + e.what(),
+            Logger::FATAL,
+            {Logger::FILESYSTEM, Logger::SAVE}
+        );
+
         return {WALLET_FILE_CORRUPTED, nullptr};
     }
 }
@@ -525,7 +567,7 @@ void WalletBackend::init()
         auto [startHeight, startTimestamp] = m_subWallets->getMinInitialSyncStart();
 
         m_walletSynchronizer = std::make_shared<WalletSynchronizer>(
-            m_daemon, 
+            m_daemon,
             startHeight,
             startTimestamp,
             m_subWallets->getPrivateViewKey(),
@@ -606,6 +648,12 @@ Error WalletBackend::unsafeSave() const
 
     if (!file)
     {
+        Logger::logger.log(
+            std::string("Wallet filename: ") + m_filename + " is invalid",
+            Logger::FATAL,
+            {Logger::FILESYSTEM, Logger::SAVE}
+        );
+
         return INVALID_WALLET_FILENAME;
     }
 
@@ -741,7 +789,7 @@ std::tuple<Error, std::string, Crypto::SecretKey> WalletBackend::addSubWallet()
 {
     return m_syncRAIIWrapper->pauseSynchronizerToRunFunction([this]() {
         /* Add the sub wallet */
-        return m_subWallets->addSubWallet(); 
+        return m_subWallets->addSubWallet();
     });
 }
 
@@ -749,11 +797,16 @@ std::tuple<Error, std::string> WalletBackend::importSubWallet(
     const Crypto::SecretKey privateSpendKey,
     const uint64_t scanHeight)
 {
+    if (Error error = validatePrivateKey(privateSpendKey); error != SUCCESS)
+    {
+        return {error, std::string()};
+    }
+
     return m_syncRAIIWrapper->pauseSynchronizerToRunFunction([&, this]() {
         /* Add the sub wallet */
         const auto [error, address] = m_subWallets->importSubWallet(
             privateSpendKey, scanHeight
-        ); 
+        );
 
         if (!error)
         {
@@ -780,11 +833,16 @@ std::tuple<Error, std::string> WalletBackend::importViewSubWallet(
     const Crypto::PublicKey publicSpendKey,
     const uint64_t scanHeight)
 {
+    if (Error error = validatePublicKey(publicSpendKey); error != SUCCESS)
+    {
+        return {error, std::string()};
+    }
+
     return m_syncRAIIWrapper->pauseSynchronizerToRunFunction([&, this]() {
         /* Add the sub wallet */
         const auto [error, address] = m_subWallets->importViewSubWallet(
             publicSpendKey, scanHeight
-        ); 
+        );
 
         if (!error)
         {
@@ -809,6 +867,13 @@ std::tuple<Error, std::string> WalletBackend::importViewSubWallet(
 
 Error WalletBackend::deleteSubWallet(const std::string address)
 {
+    const bool allowIntegratedAddresses = false;
+
+    if (Error error = validateAddresses({address}, allowIntegratedAddresses); error != SUCCESS)
+    {
+        return error;
+    }
+
     return m_syncRAIIWrapper->pauseSynchronizerToRunFunction([&, this]() {
         return m_subWallets->deleteSubWallet(address);
     });
@@ -875,6 +940,13 @@ Error WalletBackend::changePassword(const std::string newPassword)
 std::tuple<Error, Crypto::PublicKey, Crypto::SecretKey>
     WalletBackend::getSpendKeys(const std::string &address) const
 {
+    const bool allowIntegratedAddresses = false;
+
+    if (Error error = validateAddresses({address}, allowIntegratedAddresses); error != SUCCESS)
+    {
+        return {error, Crypto::PublicKey(), Crypto::SecretKey()};
+    }
+
     const auto [publicSpendKey, publicViewKey] = Utilities::addressToKeys(address);
 
     const auto [success, privateSpendKey] = m_subWallets->getPrivateSpendKey(publicSpendKey);
@@ -901,6 +973,13 @@ std::tuple<Error, std::string> WalletBackend::getMnemonicSeed() const
 std::tuple<Error, std::string> WalletBackend::getMnemonicSeedForAddress(
     const std::string &address) const
 {
+    const bool allowIntegratedAddresses = false;
+
+    if (Error error = validateAddresses({address}, allowIntegratedAddresses); error != SUCCESS)
+    {
+        return {error, std::string()}; 
+    }
+
     const auto privateViewKey = getPrivateViewKey();
     const auto [error, publicSpendKey, privateSpendKey] = getSpendKeys(address);
 
@@ -976,16 +1055,16 @@ std::tuple<uint64_t, std::string> WalletBackend::getNodeFee() const
     return m_daemon->nodeFee();
 }
 
-std::tuple<std::string, uint16_t> WalletBackend::getNodeAddress() const
+std::tuple<std::string, uint16_t, bool> WalletBackend::getNodeAddress() const
 {
     return m_daemon->nodeAddress();
 }
 
-void WalletBackend::swapNode(std::string daemonHost, uint16_t daemonPort)
+void WalletBackend::swapNode(std::string daemonHost, uint16_t daemonPort, bool daemonSSL)
 {
     m_syncRAIIWrapper->pauseSynchronizerToRunFunction([&, this]() {
         /* Swap and init the node */
-        m_daemon->swapNode(daemonHost, daemonPort);
+        m_daemon->swapNode(daemonHost, daemonPort, daemonSSL);
 
         /* Give the synchronizer the new daemon */
         m_walletSynchronizer->swapNode(m_daemon);
@@ -1002,6 +1081,11 @@ bool WalletBackend::daemonOnline() const
 std::tuple<Error, std::string> WalletBackend::getAddress(
     const Crypto::PublicKey spendKey) const
 {
+    if (Error error = validatePublicKey(spendKey); error != SUCCESS)
+    {
+        return {error, std::string()};
+    }
+
     return m_subWallets->getAddress(spendKey);
 }
 
@@ -1067,7 +1151,8 @@ Error WalletBackend::fromJSON(
     const std::string filename,
     const std::string password,
     const std::string daemonHost,
-    const uint16_t daemonPort)
+    const uint16_t daemonPort,
+    const bool daemonSSL)
 {
     if (Error error = fromJSON(j); error != SUCCESS)
     {
@@ -1077,7 +1162,7 @@ Error WalletBackend::fromJSON(
     m_filename = filename;
     m_password = password;
 
-    m_daemon = std::make_shared<Nigel>(daemonHost, daemonPort);
+    m_daemon = std::make_shared<Nigel>(daemonHost, daemonPort, daemonSSL);
 
     init();
